implement Q2(x) polynomial formulations: Q(x) = (2ax + b)^2 -kN as in Kechlibar's paper and as suggested by Till on mersenneforum.  ~85% of inputs can use kN==1 mod 8 and are ~10% faster with this change.
improvements to bucket sorting to make better use of compressstoreu.  This is only available for linux builds that are AVX512 capable, but gives around 15-20% speedup on those systems!

some documentation/comments updates

fix a bug in the calculator that caused crashes when multiple larger inputs were pasted into the interactive environment