fixes for tune and SNFS
Fixed and updated tune() function.  Now more accurate extrapolations and addition of ECM fitting over size and B1.

SNFS was looking for an exact match in the .ranges file.  It is now a divisibility check.  Also set the correct input when running SQRT so it doesn't try to find dependencies for known factors already divided out.